app/requirements.txt
APScheduler==3.8.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
Flask-APScheduler==1.12.2
freezegun==1.2.1
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.1.1
nltk==3.7
numpy==1.22.3
packaging==21.3
pandas==1.3.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycoingecko==2.2.0
pyparsing==3.0.7
pytest==7.1.1
pytest-freezegun==0.4.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.1
regex==2022.3.15
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
SQLAlchemy==1.4.25
textblob==0.17.1
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.63.0
tweet-preprocessor==0.6.0
tzlocal==2.1
urllib3==1.26.9
Werkzeug==2.0.3
xgboost==1.5.2
